Feat: Added home page with links to the different components
