+++ starters/vue3-apollo-quasar/src/App.vue
@@ -14,5 +14,5 @@ export default defineComponent({
 import { setupGraphQL } from './init';
 setupGraphQL();
 
-provide('PRODUCT_NAME', 'Vue3 | Quasar & Apollo Starter kit');
+provide('PRODUCT_NAME', 'Vue 3, Quasar & Apollo Starter kit');
 </script>
