+++ starters/vue3-apollo-quasar/src/pages/ApiExample.vue
@@ -1,5 +1,9 @@
 <template>
-  <h2>API Example</h2>
+  <div>
+    <h4 class="text-center">Vue3 Apollo Quasar Fetch Data from API</h4>
+    <q-separator color="primary" size="4px" />
+    <FetchMessage :message="message" />
+  </div>
 </template>
 
 <script lang="ts">
@@ -8,4 +12,11 @@ import { defineComponent } from 'vue';
 export default defineComponent({
   name: 'ApiExample',
 });
+</script>
+
+<script lang="ts" setup>
+import { ref } from 'vue';
+import { FetchMessage } from 'src/components';
+
+const message = ref('vue3-apollo-quasar starter.dev!');
 </script>
